[REFACTOR] Refactor structure
