+++ src/validation.ts
@@ -1,0 +1,9 @@
+export const isValidUrl = (s: string): string => {
+  try {
+    new URL(s);
+    return s;
+  } catch (e) {
+    console.error("Invalid url");
+    process.exit(128);
+  }
+};
